Ensure naming and argument validation is consistent

+++ src/Foundatio/Storage/IFileStorage.cs
@@ -14,8 +14,8 @@ namespace Foundatio.Storage {
         Task<FileSpec> GetFileInfoAsync(string path);
         Task<bool> ExistsAsync(string path);
         Task<bool> SaveFileAsync(string path, Stream stream, CancellationToken cancellationToken = default(CancellationToken));
-        Task<bool> RenameFileAsync(string path, string newpath, CancellationToken cancellationToken = default(CancellationToken));
-        Task<bool> CopyFileAsync(string path, string targetpath, CancellationToken cancellationToken = default(CancellationToken));
+        Task<bool> RenameFileAsync(string path, string newPath, CancellationToken cancellationToken = default(CancellationToken));
+        Task<bool> CopyFileAsync(string path, string targetPath, CancellationToken cancellationToken = default(CancellationToken));
         Task<bool> DeleteFileAsync(string path, CancellationToken cancellationToken = default(CancellationToken));
         Task DeleteFilesAsync(string searchPattern = null, CancellationToken cancellation = default(CancellationToken));
         Task<IEnumerable<FileSpec>> GetFileListAsync(string searchPattern = null, int? limit = null, int? skip = null, CancellationToken cancellationToken = default(CancellationToken));
@@ -36,12 +36,18 @@ namespace Foundatio.Storage {
 
     public static class FileStorageExtensions {
         public static Task<bool> SaveObjectAsync<T>(this IFileStorage storage, string path, T data, CancellationToken cancellationToken = default(CancellationToken)) {
+            if (String.IsNullOrEmpty(path))
+                throw new ArgumentNullException(nameof(path));
+
             var bytes = storage.Serializer.SerializeToBytes(data);
             return storage.SaveFileAsync(path, new MemoryStream(bytes), cancellationToken);
         }
 
         public static async Task<T> GetObjectAsync<T>(this IFileStorage storage, string path, CancellationToken cancellationToken = default(CancellationToken)) {
-            using (Stream stream = await storage.GetFileStreamAsync(path, cancellationToken).AnyContext()) {
+            if (String.IsNullOrEmpty(path))
+                throw new ArgumentNullException(nameof(path));
+
+            using (var stream = await storage.GetFileStreamAsync(path, cancellationToken).AnyContext()) {
                 if (stream != null)
                     return storage.Serializer.Deserialize<T>(stream);
             }
@@ -50,11 +56,17 @@ namespace Foundatio.Storage {
         }
 
         public static async Task DeleteFilesAsync(this IFileStorage storage, IEnumerable<FileSpec> files) {
+            if (files == null)
+                throw new ArgumentNullException(nameof(files));
+
             foreach (var file in files)
                 await storage.DeleteFileAsync(file.Path).AnyContext();
         }
 
         public static async Task<string> GetFileContentsAsync(this IFileStorage storage, string path) {
+            if (String.IsNullOrEmpty(path))
+                throw new ArgumentNullException(nameof(path));
+
             using (var stream = await storage.GetFileStreamAsync(path).AnyContext()) {
                 if (stream != null)
                     return await new StreamReader(stream).ReadToEndAsync().AnyContext();
@@ -64,11 +76,14 @@ namespace Foundatio.Storage {
         }
 
         public static async Task<byte[]> GetFileContentsRawAsync(this IFileStorage storage, string path) {
+            if (String.IsNullOrEmpty(path))
+                throw new ArgumentNullException(nameof(path));
+
             using (var stream = await storage.GetFileStreamAsync(path).AnyContext()) {
                 if (stream == null)
                     return null;
 
-                byte[] buffer = new byte[16 * 1024];
+                var buffer = new byte[16 * 1024];
                 using (var ms = new MemoryStream()) {
                     int read;
                     while ((read = await stream.ReadAsync(buffer, 0, buffer.Length).AnyContext()) > 0) {
@@ -81,6 +96,9 @@ namespace Foundatio.Storage {
         }
 
         public static Task<bool> SaveFileAsync(this IFileStorage storage, string path, string contents) {
+            if (String.IsNullOrEmpty(path))
+                throw new ArgumentNullException(nameof(path));
+
             return storage.SaveFileAsync(path, new MemoryStream(Encoding.UTF8.GetBytes(contents ?? String.Empty)));
         }
     }
